many changes

+++ java/src/main/java/cz/kul/snippets/java/example02_autoboxingunboxing/TestAutoboxingUnboxing.java
@@ -5,12 +5,12 @@ import org.junit.Test;
 
 public class TestAutoboxingUnboxing {
 
-    private Integer i = new Integer(10);
-    private Integer j = new Integer(10);
+    private Integer i = new Integer(12345);
+    private Integer j = new Integer(12345);
 
     @Test
     public void test_Integer_Integer() {
-        Assert.assertFalse(i == j); // IT is not equal !!! There is no Unboxing here.
+        Assert.assertFalse(i == j); // it is not equal !!! There is no Unboxing here.
     }
 
     @Test
@@ -23,6 +23,9 @@ public class TestAutoboxingUnboxing {
 
     @Test
     public void test_int_Integer() {
+        // you can really compare int and Integer by == operator, you does not have to manually
+        // convert Integer to int. Integer is always unboxed to primitive.
+        // see https://stackoverflow.com/questions/7672317/integer-int-allowed-in-java
         Assert.assertTrue(i.intValue() == j); // Unbox to primitive
         Assert.assertTrue(i == j.intValue()); // Unbox to primitive
     }
